Added Dockerfile and compose file.
Created a docker file for a simple static html IIS site that I can use for testing my reverse proxy.  The site has two links one to /hello and another to /goodbye.  The idea is to spin up three instances of this container, each with a different pathtext value to demonstrate going from one server to the next through the reverse proxy.